CreateMacrosFrom.jl

CreateMacrosFrom.create_macros_from

CreateMacrosFrom.@create_macros_from

Sometimes it's easier to have macros in their underlying function form (functions which take and return code). These functions can be reused in functional programming (map, reduce, filter) to build more complicated macros. This package automatically generates macros based on underlying functions. This strategy is parallel to the system of non-standard evaluation used in the Hadley-verse.

CreateMacrosFrom.create_macros_from — Method.

@create_macros_from(functions...)

Will create a macro for each of the functions.

Each function should take and return code. The new macro will have the same name.

examples

julia> using CreateMacrosFrom

julia> binary_function(a, b, c) = Expr(:call, b, a, c);

julia> reverse_binary(a, b, c) = Expr(:call, b, c, a);

julia> @create_macros_from binary_function reverse_binary;

julia> vcat(1, 2) == @binary_function 1 vcat 2
true

julia> vcat(2, 1) == @reverse_binary 1 vcat 2
true

There is no scrubbing for hygeine, so be careful. Interpolate in local functions, and don't create local variables unless you mean it.

julia> using CreateMacrosFrom

julia> set_a_to_one() = :(a = 1);

julia> @create_macros_from set_a_to_one;

julia> a = 2;

julia> @set_a_to_one;

julia> a
1

Will write a docstring for the macro pointing to the documentation of the original function.

julia> using CreateMacrosFrom

julia> "A macro to parse any function as a binary operator"
       binary_function(a, b, c) = Expr(:call, b, a, c);

julia> @create_macros_from binary_function;

julia> @doc :(@binary_function)
See documentation of [`binary_function`](@ref)

source

CreateMacrosFrom.@create_macros_from — Macro.

See documentation of create_macros_from

source